* Development 12 | Development 7 *
* Added GUI support for android
* Removed TextDisplayArea
* Added the following classes for android (in different packages to the
other ones)
  - GUI
  - GUIComponent
  - GUILabel
  - GUIPanel
  - GUIBuilder
  - GUIButtonBuilder
  - GUICheckBoxBuilder
  - GUILoadingBarBuilder
  - GUIButton
  - GUIImageButton
  - GUIRenderableButton
  - GUIButtonEvent
  - GUICheckBoxEvent
  - GUIButtonListener
  - GUICheckBoxListener
  - GUICheckBox
  - GUIImageCheckBox
  - GUIRenderableCheckBox
  - GUIFont
  - GUIImageLoadingBar
  - GUIRenderableLoadingBar
  - GUITextDisplayArea
  - FontUtils
- Added getAndroidBounds() method in Entity2D
- Added the colour orange
- Updated the version in the settings
- Found an issue: If you set a variable to a colour constant then change
its a,r,g,b value the colour constant will change with the variable

+++ MyEngine-Master/src/org/simplecorporation/myengine/core/gui/font/GUIFont.java
@@ -28,18 +28,18 @@ public class GUIFont {
 	public TrueTypeFont ttfFont;
 	
 	/* The font size */
-	public float size;
+	public double size;
 	
 	/* The colour */
 	public Colour colour;
 	
 	/* The constructor of the font */
-	public GUIFont(Font font , Colour colour , float size) {
+	public GUIFont(Font font , Colour colour , double size) {
 		//Set the font
 		this.font = font;
 		
 		if (Settings.Video.OpenGL)
-			ttfFont = new TrueTypeFont(this.font.deriveFont(size) , Settings.Video.AntiAliasing);
+			ttfFont = new TrueTypeFont(this.font.deriveFont((float)size) , Settings.Video.AntiAliasing);
 		
 		//Set the colour and size
 		this.colour = colour;
@@ -58,7 +58,7 @@ public class GUIFont {
 			//Set the colour
 			BasicRenderer.setColour(this.colour);
 			//Set the font
-			JavaWindow.g2d.setFont(this.font.deriveFont(size));
+			JavaWindow.g2d.setFont(this.font.deriveFont((float)size));
 			//Render the text
 			JavaWindow.g2d.drawString(text , (float) x , (float) y);
 		}
@@ -69,7 +69,7 @@ public class GUIFont {
 		if (! Settings.Android && Settings.Video.OpenGL) {
 			return ttfFont.getWidth(text);
 		} else if (! Settings.Android && ! Settings.Video.OpenGL) {
-			FontMetrics metrics = JavaWindow.g2d.getFontMetrics(font.deriveFont(size));
+			FontMetrics metrics = JavaWindow.g2d.getFontMetrics(font.deriveFont((float)size));
 			return metrics.getStringBounds(text , JavaWindow.g2d).getWidth();
 		} else
 			return 0;
@@ -80,7 +80,7 @@ public class GUIFont {
 		if (! Settings.Android && Settings.Video.OpenGL) {
 			return ttfFont.getHeight(text);
 		} else if (! Settings.Android && ! Settings.Video.OpenGL) {
-			FontMetrics metrics = JavaWindow.g2d.getFontMetrics(font.deriveFont(size));
+			FontMetrics metrics = JavaWindow.g2d.getFontMetrics(font.deriveFont((float)size));
 			return metrics.getStringBounds(text , JavaWindow.g2d).getHeight();
 		} else
 			return 0;
